Avoid intellij crashes (#20)
I've noticed a few crash cases when developing jsonnet locally:
- sufficiently gnarly jsonnet can cause the lookup to infinite loop and
blow the stack
- IntelliJ doesn't always consistently update the version index by 1,
leading to a crash on the diff update. typically violations are just
skipping a single version number and as far as I can tell, the LSP state
would have been fine if it had just accepted the new update -- these
cases appear to have happened during git operations in which IntelliJ is
sending the whole file as an update anyway.